renames pki check-config decorator to check-issuers
