+++ ui/lib/pki/addon/routes/overview.js
@@ -5,7 +5,7 @@
 
 import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
-import { withConfig } from 'pki/decorators/check-config';
+import { withConfig } from 'pki/decorators/check-issuers';
 import { hash } from 'rsvp';
 
 export const PKI_DEFAULT_EMPTY_STATE_MSG =
